fix(api): Optimize fetching latest deploys in `ProjectSummarySerializer`
For projects with many `ReleaseProjectEnvironment` rows we end up loading a lot of models into
memory. We then load the latest `Deploy` for each row, which can be thousands in extreme cases.

Convert this to happen entirely in postgres via overuse of `OVER`. We also artificially limit the
number of `ReleaseProjectEnvironment` rows we check, since it's likely only recent rows will be
relevant.